+++ src/main/java/org/opensearch/searchrelevance/transport/searchConfiguration/PutSearchConfigurationRequest.java
@@ -19,12 +19,14 @@ public class PutSearchConfigurationRequest extends ActionRequest {
     private final String index;
     private final String queryBody;
     private final String searchPipeline;
+    private final String description;
 
-    public PutSearchConfigurationRequest(String name, String index, String queryBody, String searchPipeline) {
+    public PutSearchConfigurationRequest(String name, String index, String queryBody, String searchPipeline, String description) {
         this.name = name;
         this.index = index;
         this.queryBody = queryBody;
         this.searchPipeline = searchPipeline;
+        this.description = description;
     }
 
     public PutSearchConfigurationRequest(StreamInput in) throws IOException {
@@ -33,6 +35,7 @@ public class PutSearchConfigurationRequest extends ActionRequest {
         this.index = in.readString();
         this.queryBody = in.readString();
         this.searchPipeline = in.readOptionalString();
+        this.description = in.readOptionalString();
     }
 
     @Override
@@ -42,6 +45,7 @@ public class PutSearchConfigurationRequest extends ActionRequest {
         out.writeString(index);
         out.writeString(queryBody);
         out.writeOptionalString(searchPipeline);
+        out.writeOptionalString(description);
     }
 
     public String getName() {
@@ -58,6 +62,10 @@ public class PutSearchConfigurationRequest extends ActionRequest {
 
     public String getSearchPipeline() {
         return searchPipeline;
+    }
+
+    public String getDescription() {
+        return description;
     }
 
     @Override